refactored how highlight is handled
